[FIX] 크래프팅 리스트 창 업데이트 안되던 버그 수정
쓰레기통 슬롯에 아이템을 덮어써서 아이템이 사라져도, 크래프팅창에 시각적 반영이 안 되던 버그를 수정

+++ Source/AWeek/Components/AWeekInventoryComponent.h
@@ -210,5 +210,5 @@ protected:
 	virtual int AddNewItem(UAWeekItemBase* Item, const int32 AmountToAdd, int32 TargetIndex = -1);
 private:
 	void SetItemQuantity(FAWeekInventorySlotData& ItemSlot, const int32 Quantity);
-	void UpdateInventoryTotalWeight(int32 DeltaWeight);
+	void UpdateInventoryTotalWeight(float DeltaWeight);
 };
